refactor[#40] : PopularPostWidget 에서 use(promise) 로 데이터를 읽도록 변경
